backend_simple/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pydantic[email]>=2.9.0
email-validator>=2.1.0
